+++ sphinx_simplepdf/__init__.py
@@ -1,0 +1,6 @@
+from .directives.ifbuilder import IfBuilderDirective, IfBuilder
+
+
+def setup(app):
+    app.add_node(IfBuilder)
+    app.add_directive('if-builder', IfBuilderDirective)